Drop support for Lighttpd https://bugs.webkit.org/show_bug.cgi?id=268292
Reviewed by Alexey Proskuryakov.

Every port nowadays has `_use_apache` returning True, so this is now all
dead code.

* Tools/Scripts/webkitpy/layout_tests/servers/http_server.py: Removed.
* Tools/Scripts/webkitpy/layout_tests/servers/http_server_base.py:
* Tools/Scripts/webkitpy/layout_tests/servers/lighttpd.conf: Removed.
* Tools/Scripts/webkitpy/layout_tests/servers/websocket_server.py:
(PyWebSocket):
(PyWebSocket.__init__):
(PyWebSocket._stop_running_server):
(PyWebSocket._check_and_kill):
(PyWebSocket._is_server_running_on_all_ports):
* Tools/Scripts/webkitpy/layout_tests/servers/websocket_server_unittest.py: Renamed from Tools/Scripts/webkitpy/layout_tests/servers/http_server_unittest.py.
(TestWebsocketServer):
(TestWebsocketServer.test_start_cmd):
(TestWebsocketServer.test_win32_start_and_stop):
(TestWebsocketServer.test_win32_start_and_stop.wait_for_action):
(TestWebsocketServer.test_win32_start_and_stop.mock_returns):
(TestWebsocketServer.test_win32_start_and_stop.mock_returns.return_value_thunk):
* Tools/Scripts/webkitpy/port/base.py:
(Port.check_httpd):
(Port.start_http_server):
(Port):
(Port.path_to_api_test_binaries):
(Port._uses_apache): Deleted.
(Port._path_to_lighttpd): Deleted.
(Port._path_to_lighttpd_modules): Deleted.
* Tools/Scripts/webkitpy/port/test.py:
* Tools/Scripts/webkitpy/port/win.py:
(WinPort):
(WinPort._path_to_apache):
(WinPort._uses_apache): Deleted.
(WinPort._path_to_lighttpd): Deleted.
(WinPort._path_to_lighttpd_modules): Deleted.
* Tools/Scripts/webkitpy/test/main.py:
(main):

Canonical link: https://commits.webkit.org/273698@main
