+++ Tools/Scripts/webkitpy/test/main.py
@@ -215,8 +215,8 @@ def main():
         "webkitpy.layout_tests.servers.apache_http_server_unittest",
         "webkitpy.layout_tests.servers.http_server_base_unittest",
         "webkitpy.layout_tests.servers.http_server_integrationtest",
-        "webkitpy.layout_tests.servers.http_server_unittest",
         "webkitpy.layout_tests.servers.web_platform_test_server_unittest",
+        "webkitpy.layout_tests.servers.websocket_server_unittest",
         "webkitpy.layout_tests.views.buildbot_results_unittest",
         "webkitpy.layout_tests.views.metered_stream_unittest",
         "webkitpy.layout_tests.views.printing_unittest",
